+++ src/main/java/com/fateczl/SistemaGerenciamentoWEB/controller/FornecedoresController.java
@@ -16,6 +16,7 @@ import org.springframework.web.servlet.ModelAndView;
 import com.fateczl.SistemaGerenciamentoWEB.model.Endereco;
 import com.fateczl.SistemaGerenciamentoWEB.model.Fornecedor;
 import com.fateczl.SistemaGerenciamentoWEB.persistence.InterfaceDAO.FornecedoresDAO;
+import com.fateczl.SistemaGerenciamentoWEB.persistence.InterfaceDAO.LoginDAO;
 
 @Controller
 public class FornecedoresController {
@@ -23,6 +24,8 @@ public class FornecedoresController {
 	@Autowired
 	FornecedoresDAO fDAO;
 	
+	@Autowired
+	LoginDAO lDAO;
 	private static int id_fornecedor;
 	
 	@RequestMapping(name="Fornecedores", value="/Fornecedores", method=RequestMethod.GET)
@@ -31,14 +34,19 @@ public class FornecedoresController {
 		
 		List<Fornecedor> listaFornecedores = new ArrayList<Fornecedor>();
 		try {
-			listaFornecedores = fDAO.listaFornecedor();
+			if(lDAO.verificarAcesso().equals("Administrador")) {
+				listaFornecedores = fDAO.listaFornecedor();
+				model.addAttribute("listaFornecedores", listaFornecedores);
+				return new ModelAndView("Fornecedores");
+			}else {
+				erro = "Acesso não autorizado";
+				model.addAttribute("erro", erro);
+				return new ModelAndView("Fornecedores");
+			}
 		} catch (ClassNotFoundException | SQLException e) {
 			erro = e.getMessage();
-		} finally {
-			model.addAttribute("erro", erro);
-			model.addAttribute("listaFornecedores", listaFornecedores);
-		}
-		return new ModelAndView();
+		} 
+		return new ModelAndView("Fornecedores");
 	}
 	@RequestMapping(name="Fornecedores", value="/Fornecedores", method=RequestMethod.POST)
 	public ModelAndView fornecedores(ModelMap model, @RequestParam Map<String,String> param) {
@@ -48,28 +56,33 @@ public class FornecedoresController {
 		List<Fornecedor> listaFornecedores = new ArrayList<Fornecedor>();
 		
 		try {
-			if(botaoInput.isEmpty()) {
-				listaFornecedores = fDAO.listaFornecedor();
-			}else{
-				listaFornecedores = fDAO.pesquisarFornecedoresPorNome(botaoInput);
-				if(listaFornecedores.isEmpty()) {
+			if(lDAO.verificarAcesso().equals("Administrador")) {
+				if(botaoInput.isEmpty()) {
 					listaFornecedores = fDAO.listaFornecedor();
-					model.addAttribute("erro", erro);
-					model.addAttribute("listaFornecedores", listaFornecedores);
-					
-				}else {	
-					model.addAttribute("erro", erro);
-					model.addAttribute("listaFornecedores", listaFornecedores);
-					return new ModelAndView("Fornecedores");
+				}else{
+					listaFornecedores = fDAO.pesquisarFornecedoresPorNome(botaoInput);
+					if(listaFornecedores.isEmpty()) {
+						listaFornecedores = fDAO.listaFornecedor();
+						model.addAttribute("erro", erro);
+						model.addAttribute("listaFornecedores", listaFornecedores);
+						
+					}else {	
+						model.addAttribute("erro", erro);
+						model.addAttribute("listaFornecedores", listaFornecedores);
+						return new ModelAndView("Fornecedores");
+					}
 				}
+				if(botaoEditar != null && !botaoEditar.isEmpty()) {
+					id_fornecedor = Integer.parseInt(botaoEditar);
+					fornecedorEditar(model);
+					model.addAttribute("listaFornecedores", listaFornecedores);
+					return new ModelAndView("FornecedorEditar");
+				}
+			}else {
+				erro = "Acesso não autorizado";
+				model.addAttribute("erro", erro);
+				return new ModelAndView("Fornecedores");
 			}
-			if(botaoEditar != null && !botaoEditar.isEmpty()) {
-				id_fornecedor = Integer.parseInt(botaoEditar);
-				fornecedorEditar(model);
-				model.addAttribute("listaFornecedores", listaFornecedores);
-				return new ModelAndView("FornecedorEditar");
-			}
-			
 		} catch (ClassNotFoundException | SQLException e) {
 			erro = e.getMessage();
 		} finally {
@@ -80,6 +93,18 @@ public class FornecedoresController {
 	}
 	@RequestMapping(name="FornecedorAdicionar", value="/FornecedorAdicionar", method=RequestMethod.GET)
 	public ModelAndView init(ModelMap model) {
+		String erro = "";
+		try {
+			if(lDAO.verificarAcesso().equals("Administrador")) {
+				return new ModelAndView();
+			}else {
+				erro = "Acesso não autorizado";
+				model.addAttribute("erro", erro);
+				return new ModelAndView("FornecedorAdicionar");
+			}
+		} catch (ClassNotFoundException | SQLException e) {
+			e.printStackTrace();
+		}
 		return new ModelAndView();
 	}
 	@RequestMapping(name="FornecedorAdicionar", value="/FornecedorAdicionar", method=RequestMethod.POST)
@@ -87,32 +112,36 @@ public class FornecedoresController {
 		String botaoAdicionar = param.get("botaoAdicionar");
 		String erro = "";
 		try {
-			if (!botaoAdicionar.isEmpty()) {
-				Fornecedor fornecedor = new Fornecedor();
-				
-				fornecedor.setRazaoSocial(param.get("razaoSocial"));
-				fornecedor.setCnpj(param.get("cnpj"));
-				fornecedor.setTelefone(param.get("telefone"));
-				fornecedor.setInscricaoEstadual(param.get("inscricaoEstadual"));
-				
-				Endereco end = new Endereco();
-				end.setCep(param.get("CEP"));
-				end.setCidade(param.get("Cidade"));
-				end.setComplemento(param.get("Complemento"));
-				end.setEstado(param.get("Estado"));
-				end.setLogradouro(param.get("Logradouro"));
-				end.setNumero(Integer.parseInt(param.get("Numero")));
-		
-				if (fDAO.verificarDuplicidade(fornecedor.getCnpj())) {
-					fDAO.adicionarFornecedor(fornecedor, end);
-				} else {
-					System.out.println("duplicado");
+			if(lDAO.verificarAcesso().equals("Administrador")) {
+				if (!botaoAdicionar.isEmpty()) {
+					Fornecedor fornecedor = new Fornecedor();
+					
+					fornecedor.setRazaoSocial(param.get("razaoSocial"));
+					fornecedor.setCnpj(param.get("cnpj"));
+					fornecedor.setTelefone(param.get("telefone"));
+					fornecedor.setInscricaoEstadual(param.get("inscricaoEstadual"));
+					
+					Endereco end = new Endereco();
+					end.setCep(param.get("CEP"));
+					end.setCidade(param.get("Cidade"));
+					end.setComplemento(param.get("Complemento"));
+					end.setEstado(param.get("Estado"));
+					end.setLogradouro(param.get("Logradouro"));
+					end.setNumero(Integer.parseInt(param.get("Numero")));
+			
+					if (fDAO.verificarDuplicidade(fornecedor.getCnpj())) {
+						fDAO.adicionarFornecedor(fornecedor, end);
+					} else {
+						System.out.println("duplicado");
+					}
 				}
+			}else {
+				erro = "Acesso não autorizado";
+				model.addAttribute("erro", erro);
+				return new ModelAndView("FornecedorAdicionar");
 			}
 		} catch (ClassNotFoundException | SQLException e) {
 			erro = e.getMessage();
-		}finally {
-			model.addAttribute("erro", erro);
 		}
 		return new ModelAndView();
 	}
@@ -120,15 +149,21 @@ public class FornecedoresController {
 	public ModelAndView fornecedorEditar(ModelMap model) {
 		Fornecedor fornecedor = new Fornecedor();
 		Endereco endereco = new Endereco();
-		
+		String erro = "";
 		try {
-			fornecedor = fDAO.buscarFornecedorPorId(id_fornecedor);
-			endereco = fDAO.buscarEnderecoFornecedorPorId(id_fornecedor);
+			if(lDAO.verificarAcesso().equals("Administrador")) {
+				fornecedor = fDAO.buscarFornecedorPorId(id_fornecedor);
+				endereco = fDAO.buscarEnderecoFornecedorPorId(id_fornecedor);
+				model.addAttribute("fornecedor", fornecedor);
+				model.addAttribute("endereco", endereco);
+				return new ModelAndView("FornecedorEditar");
+			}else {
+				erro = "Acesso não autorizado";
+				model.addAttribute("erro", erro);
+				return new ModelAndView("FornecedorEditar");
+			}
 		}catch(ClassNotFoundException | SQLException e) {
 			e.printStackTrace();
-		}finally {
-			model.addAttribute("fornecedor", fornecedor);
-			model.addAttribute("endereco", endereco);
 		}
 		return new ModelAndView("FornecedorEditar");
 	}
@@ -139,28 +174,34 @@ public class FornecedoresController {
 		String erro = "";
 		
 		try {
-			Fornecedor fornecedor = new Fornecedor();
-			Endereco endereco = new Endereco();
+			if(lDAO.verificarAcesso().equals("Administrador")) {
+				Fornecedor fornecedor = new Fornecedor();
+				Endereco endereco = new Endereco();
+				
+				fornecedor.setId(Integer.parseInt(param.get("Id")));
+				fornecedor.setRazaoSocial(param.get("RazaoSocial"));
+				fornecedor.setCnpj(param.get("Cnpj"));
+				fornecedor.setInscricaoEstadual(param.get("InscricaoEstadual"));
+				fornecedor.setTelefone(param.get("Telefone"));
+				
+				endereco.setCep(param.get("CEP"));
+				endereco.setLogradouro(param.get("Logradouro"));
+				endereco.setNumero(Integer.parseInt(param.get("Numero")));
+				endereco.setComplemento(param.get("Complemento"));
+				endereco.setCidade(param.get("Cidade"));
+				endereco.setEstado(param.get("Estado"));
 			
-			fornecedor.setId(Integer.parseInt(param.get("Id")));
-			fornecedor.setRazaoSocial(param.get("RazaoSocial"));
-			fornecedor.setCnpj(param.get("Cnpj"));
-			fornecedor.setInscricaoEstadual(param.get("InscricaoEstadual"));
-			fornecedor.setTelefone(param.get("Telefone"));
-			
-			endereco.setCep(param.get("CEP"));
-			endereco.setLogradouro(param.get("Logradouro"));
-			endereco.setNumero(Integer.parseInt(param.get("Numero")));
-			endereco.setComplemento(param.get("Complemento"));
-			endereco.setCidade(param.get("Cidade"));
-			endereco.setEstado(param.get("Estado"));
-		
-			if(botaoSalvar != null && !botaoSalvar.isEmpty()) {
-				fDAO.editarFornecedor(fornecedor, endereco);
-				return new ModelAndView("Fornecedores");
-			}
-			if(botaoExcluir != null && !botaoExcluir.isEmpty()) {
-				fDAO.excluirFornecedorPorId(fornecedor.getId());
+				if(botaoSalvar != null && !botaoSalvar.isEmpty()) {
+					fDAO.editarFornecedor(fornecedor, endereco);
+					return new ModelAndView("Fornecedores");
+				}
+				if(botaoExcluir != null && !botaoExcluir.isEmpty()) {
+					fDAO.excluirFornecedorPorId(fornecedor.getId());
+					return new ModelAndView("FornecedorEditar");
+				}
+			}else {
+				erro = "Acesso não autorizado";
+				model.addAttribute("erro", erro);
 				return new ModelAndView("FornecedorEditar");
 			}
 		}catch (ClassNotFoundException | SQLException e) {
