+++ src/main/java/com/fateczl/SistemaGerenciamentoWEB/controller/GerenciarUsuariosController.java
@@ -15,6 +15,7 @@ import org.springframework.web.servlet.ModelAndView;
 
 import com.fateczl.SistemaGerenciamentoWEB.model.Usuario;
 import com.fateczl.SistemaGerenciamentoWEB.persistence.InterfaceDAO.GerenciarUsuariosDAO;
+import com.fateczl.SistemaGerenciamentoWEB.persistence.InterfaceDAO.LoginDAO;
 
 @Controller
 public class GerenciarUsuariosController {
@@ -22,6 +23,8 @@ public class GerenciarUsuariosController {
 	@Autowired
 	GerenciarUsuariosDAO guDAO;
 	
+	@Autowired
+	LoginDAO lDAO;
 	private static int id_usuario;
 	
 	@RequestMapping(name="GerenciarUsuarios", value="/GerenciarUsuarios", method=RequestMethod.GET)
@@ -29,14 +32,19 @@ public class GerenciarUsuariosController {
 		String erro="";
 		List<Usuario> listarUsuario = new ArrayList<Usuario>();
 		try {
-			listarUsuario = guDAO.listaUsuario();
+			if(lDAO.verificarAcesso().equals("Administrador")) {
+				listarUsuario = guDAO.listaUsuario();
+				model.addAttribute("listarUsuario", listarUsuario);
+				return new ModelAndView("GerenciarUsuarios");
+			}else {
+				erro = "Acesso não autorizado";
+				model.addAttribute("erro", erro);
+				return new ModelAndView("GerenciarUsuarios");
+			}
 		}catch (ClassNotFoundException | SQLException e) {
 			erro = e.getMessage();
-		}finally {
-			model.addAttribute("erro", erro);
-			model.addAttribute("listarUsuario", listarUsuario);
 		}
-		return new ModelAndView();
+		return new ModelAndView("GerenciarUsuarios");
 	}
 	@RequestMapping(name="GerenciarUsuarios", value="/GerenciarUsuarios", method=RequestMethod.POST)
 	public ModelAndView listarUsuario(ModelMap model, @RequestParam Map<String, String> param) {
@@ -46,41 +54,60 @@ public class GerenciarUsuariosController {
 		
 		List<Usuario> listaUsuario = new ArrayList<Usuario>();
 		try {
-			if(botaoInput.isEmpty()) {
-				listaUsuario = guDAO.listaUsuario();
-			}else {
-				listaUsuario = guDAO.pesquisarUsuarioPorNome(botaoInput);
-				if(listaUsuario.isEmpty()) {
+			if(lDAO.verificarAcesso().equals("Administrador")) {
+				if(botaoInput.isEmpty()) {
 					listaUsuario = guDAO.listaUsuario();
-					model.addAttribute("listarUsuario", listaUsuario);
-					return new ModelAndView("GerenciarUsuarios");
 				}else {
-					model.addAttribute("listarUsuario", listaUsuario);
-					return new ModelAndView("GerenciarUsuarios");
+					listaUsuario = guDAO.pesquisarUsuarioPorNome(botaoInput);
+					if(listaUsuario.isEmpty()) {
+						listaUsuario = guDAO.listaUsuario();
+						model.addAttribute("listarUsuario", listaUsuario);
+						return new ModelAndView("GerenciarUsuarios");
+					}else {
+						model.addAttribute("listarUsuario", listaUsuario);
+						return new ModelAndView("GerenciarUsuarios");
+					}
 				}
-			}
-			if(botaoEditar != null && !botaoEditar.isEmpty()) {
-				id_usuario = Integer.parseInt(param.get("botaoEditar"));
-				editarUsuario(model);
+				if(botaoEditar != null && !botaoEditar.isEmpty()) {
+					id_usuario = Integer.parseInt(param.get("botaoEditar"));
+					editarUsuario(model);
+					model.addAttribute("erro", erro);
+					model.addAttribute("listaUsuario", listaUsuario);
+					return new ModelAndView("GerenciarUsuariosEditar");
+				}
+			}else {
+				erro = "Acesso não autorizado";
 				model.addAttribute("erro", erro);
-				model.addAttribute("listaUsuario", listaUsuario);
-				return new ModelAndView("GerenciarUsuariosEditar");
+				return new ModelAndView("GerenciarUsuarios");
 			}
 		} catch (ClassNotFoundException | SQLException e) {
 			erro = e.getMessage();
 		}finally{
 			model.addAttribute("listarUsuario", listaUsuario);
 		}
-		return new ModelAndView();
+		return new ModelAndView("GerenciarUsuarios");
 	}
 	
 	@RequestMapping(name="GerenciarUsuariosAdicionar", value="/GerenciarUsuariosAdicionar", method=RequestMethod.GET)
 	public ModelAndView adicionarUsuario(ModelMap model) {
+		String erro = "";
+		try {
+			if(lDAO.verificarAcesso().equals("Administrador")) {
+				return new ModelAndView("GerenciarUsuariosAdicionar");
+			}else {
+				erro = "Acesso não autorizado";
+				model.addAttribute("erro", erro);
+				return new ModelAndView("GerenciarUsuariosAdicionar");
+			}
+		} catch (ClassNotFoundException | SQLException e) {
+			e.printStackTrace();
+		}
 		return new ModelAndView();
 	}
 	@RequestMapping(name="GerenciarUsuariosAdicionar", value="/GerenciarUsuariosAdicionar", method=RequestMethod.POST)
 	public ModelAndView adicionarUsuario(ModelMap model, @RequestParam Map<String, String> param) {
 		Usuario usuario = new Usuario();
+		String erro = "";
 		
 		usuario.setNome(param.get("nome"));
 		usuario.setLogin(param.get("login_usuario"));
@@ -98,10 +125,18 @@ public class GerenciarUsuariosController {
 		}
 		
 		try {
-			if(guDAO.verificarDuplicidade(usuario.getLogin(), usuario.getEmail())) {
-				guDAO.adicionarUsuario(usuario);
+			if(lDAO.verificarAcesso().equals("Administrador")) {
+				if(guDAO.verificarDuplicidade(usuario.getLogin(), usuario.getEmail())) {
+					guDAO.adicionarUsuario(usuario);
+				}else {
+					erro = "Usuario já cadastrado";
+					model.addAttribute("erro", erro);
+					return new ModelAndView("GerenciarUsuariosAdicionar");
+				}
 			}else {
-				System.out.println("Duplicado");
+				erro = "Acesso não autorizado";
+				model.addAttribute("erro", erro);
+				return new ModelAndView("GerenciarUsuariosAdicionar");
 			}
 		} catch (ClassNotFoundException | SQLException e) {
 			e.printStackTrace();
@@ -112,13 +147,19 @@ public class GerenciarUsuariosController {
 	@RequestMapping(name="GerenciarUsuariosEditar", value="/GerenciarUsuariosEditar", method=RequestMethod.GET)
 	public ModelAndView editarUsuario(ModelMap model) {
 		Usuario usuario = new Usuario();
-		
+		String erro = "";
 		try {
-			usuario = guDAO.pesquisarUsuarioPorId(id_usuario);
+			if(lDAO.verificarAcesso().equals("Administrador")) {
+				usuario = guDAO.pesquisarUsuarioPorId(id_usuario);
+				model.addAttribute("usuario", usuario);
+				return new ModelAndView("GerenciarUsuariosEditar");
+			}else {
+				erro = "Acesso não autorizado";
+				model.addAttribute("erro", erro);
+				return new ModelAndView("GerenciarUsuariosEditar");
+			}
 		}catch (ClassNotFoundException | SQLException e) {
 			e.printStackTrace();
-		}finally {
-			model.addAttribute("usuario", usuario);
 		}
 		return new ModelAndView("GerenciarUsuariosEditar");
 	}
@@ -129,30 +170,36 @@ public class GerenciarUsuariosController {
 		String erro = "";
 		
 		try {
-			Usuario usuario = new Usuario();
-			
-			usuario.setId(Integer.parseInt(param.get("id")));
-			usuario.setNome(param.get("nome"));
-			usuario.setLogin(param.get("email"));
-			
-			switch (param.get("tipoDeUsuario")) {
-			case "Administrador":
-				usuario.setTipo_usuario("Administrador");
-				break;
-			case "Estoquista":
-				usuario.setTipo_usuario("Estoquista");
-				break;
-			case "Vendedor":
-				usuario.setTipo_usuario("Vendedor");
-			}
-			if(botaoSalvar != null && !botaoSalvar.isEmpty()) {
-				guDAO.editarUsuarioPorId(usuario);
-				model.addAttribute("usuario", usuario);
-				return new ModelAndView("GerenciarUsuariosEditar");
-			}
-			if(botaoExcluir != null && !botaoExcluir.isEmpty()) {
-				guDAO.excluirUsuarioPorId(usuario.getId());
-				model.addAttribute("usuario", usuario);
+			if(lDAO.verificarAcesso().equals("Administrador")) {
+				Usuario usuario = new Usuario();
+				
+				usuario.setId(Integer.parseInt(param.get("id")));
+				usuario.setNome(param.get("nome"));
+				usuario.setLogin(param.get("email"));
+				
+				switch (param.get("tipoDeUsuario")) {
+				case "Administrador":
+					usuario.setTipo_usuario("Administrador");
+					break;
+				case "Estoquista":
+					usuario.setTipo_usuario("Estoquista");
+					break;
+				case "Vendedor":
+					usuario.setTipo_usuario("Vendedor");
+				}
+				if(botaoSalvar != null && !botaoSalvar.isEmpty()) {
+					guDAO.editarUsuarioPorId(usuario);
+					model.addAttribute("usuario", usuario);
+					return new ModelAndView("GerenciarUsuariosEditar");
+				}
+				if(botaoExcluir != null && !botaoExcluir.isEmpty()) {
+					guDAO.excluirUsuarioPorId(usuario.getId());
+					model.addAttribute("usuario", usuario);
+					return new ModelAndView("GerenciarUsuariosEditar");
+				}
+			}else {
+				erro = "Acesso não autorizado";
+				model.addAttribute("erro", erro);
 				return new ModelAndView("GerenciarUsuariosEditar");
 			}
 		}catch (ClassNotFoundException | SQLException e) {
